Time: 0 ms (100%), Space: 23.5 MB (83.21%) - Medium

+++ 0229-majority-element-ii/0229-majority-element-ii.cpp
@@ -60,7 +60,7 @@ public:
         // Count candidates appearences
         count_a = 0;
         count_b = 0;
-        size_t n = nums.size();
+        int n = nums.size();
 
         for (int num : nums) {
             if (num == candidate_a) {
@@ -82,7 +82,5 @@ public:
 
         //might be empty if no candidate is valid
         return result;
-
     }
 };
-
